export and centralize database calls
