+++ screens/MoreScreen.js
@@ -8,8 +8,6 @@ import {
   Alert,
   Button
 } from "react-native";
-import firebase from "firebase";
-import dbh from "../firebase";
 import colors from "../style/colors.js";
 import Header from "../components/header";
 import Footer from "../components/footer";
