+++ screens/RemindersScreen.js
@@ -7,8 +7,6 @@ import {
   TouchableOpacity,
   Alert,
 } from "react-native";
-import firebase from "firebase";
-import dbh from "../firebase";
 import colors from "../style/colors.js";
 import Header from "../components/header";
 import Footer from "../components/footer";
